+++ commons-microservicios/src/main/java/com/ideasconnections/microservicios/commons/controllers/CommonController.java
@@ -1,0 +1,42 @@
+package com.ideasconnections.microservicios.commons.controllers;
+
+import com.ideasconnections.microservicios.commons.services.CommonService;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
+
+public class CommonController<E,S extends CommonService<E>> {
+
+    @Autowired
+    protected S service;
+
+    @GetMapping
+    public ResponseEntity<?> listar() {
+        return ResponseEntity.ok().body(service.findAll());
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<?> ver(@PathVariable Long id) {
+        Optional<E> optionalE = service.findById(id);
+        if (!optionalE.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(optionalE.get());
+    }
+
+    @PostMapping
+    public ResponseEntity<?> crear(@RequestBody E entity) {
+        E entityDb = service.save(entity);
+        return ResponseEntity.status(HttpStatus.CREATED).body(entityDb);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> eliminar(@PathVariable Long id) {
+        service.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+}
